Initial Commit
Script re-formats edx course structures for use in R. Creates file of course temporal metadata, course identifiers. Second, the scripts creates an dataframe of all modules found in the structure, which sets the record order to the order that the course modules are sequenced. It assigns tree levels to each module (e.g. 0 for course, 4 for content modules), and identifies the parent chapters, sequence, and vertical parents of all children modules.